Authorize code - step by step to display the device code
In the authorize code method the identification & authorization is completed in one step which does not allow displaying the device code. Without the device code the user can not verify if the authentication is as intended (UX Issue).
This commit creates two methods that split the authorize code in two. This allows first getting the response with the device code and then going for the authentication afterwards.

                 final codeResponse =
                        await OpenIdConnect.authorizeDeviceGetDeviceCodeResponse(request: request);
                    setState(() {
                      deviceIdentification = codeResponse.userCode;
                    });

                    final response =
                        await OpenIdConnect.authorizeDeviceCompleteDeviceCodeResponseRequest(
                      request: request,
                      codeResponse: codeResponse,
                    );